DRYing up the code
Big chunks divided into classes and modules to simplify the work. COMP mode in progress
