+++ bill/comp_mode.py
@@ -2,17 +2,7 @@
 #
 # This is implementation of COMP mode :D
 # Computer will play and pocket the balls available on screen, with some factor of intellengece (manually set by me)
-# 
-# How to use it ??
-# Ball is placed on screen, click left button of mouse to have a stick on screen
-# You can feel the streching of stick if you take the mouse pointer away from the stick end point
-# Press right button of mouse to hit the stick :D
-# Ball will move in appropriate direction :D
-# 
-# Working billiards game, 
-# Balls will disappear after being pocketed :D
-# 6 pockets are added :D
-# After all balls are pocketed, then game will stop, you can quit or continue the game :D
+
 # ----------------------------------------------------------------------------------------------------------
 import pygame
 
@@ -20,629 +10,20 @@ import pygame
 import random
 from math import *
 pygame.init()
-
-white = (255, 255, 255)
-black = (0, 0, 0)
-blue = (0, 0, 255)
-red = (255, 0, 0)
-GREEN = (0,140,0)
-FPS = 100
-dispHeight = 500
-dispWidth = 800
-font = pygame.font.SysFont(None, 25)
-
-# gameDisplay = pygame.display.set_mode((dispHeight,dispWidth))
-gameDisplay = pygame.display.set_mode((dispWidth, dispHeight))
-pygame.display.set_caption('pyBilliard')
-clock = pygame.time.Clock()
+#============
+# Functions are written to some py file :D
+from modules.all_functions import *
+from classes.env_variables import *
+from classes.Balls_class import Balls
+#============
 
 pygame.display.update()
 # pygame.display.flip()
 
-class Balls:
-    def __init__(self, (x,y), size, thickness=0, color=(0,0,255), pocket_size = 0):
-        self.x = x
-        self.y = y
-        self.size = size
-        self.thickness = thickness              # Default :D
-        self.color = color
-        self.c1 = color[0]
-        self.c2 = color[1]
-        self.c3 = color[2]
-        self.speed = 1
-        self.angle = 0                           # This is in radian :D
-        if pocket_size == 0:
-            pocket_size = 2*self.size
-
-        self.pocket_size = pocket_size
-
-        self.pocketed = self.is_pocketed(self.pocket_size)      # Dpending on its initialization location :D
-        self.in_line_with_white_ball = 0        # Default is zero, but will be changed afterwards :D
-        self.ok_to_hit = 0                      # This will be made equal to 1 for white_ball in line with this ball
-
-    def disp(self):
-        pygame.draw.circle(gameDisplay, self.color, (self.x, self.y), self.size, self.thickness)
-        pygame.display.update()
-
-    def boundary(self):
-        if self.x > dispWidth - self.size:
-            self.x = 2*(dispWidth - self.size) - self.x
-            self.angle = - self.angle
-
-        elif self.x < self.size:
-            self.x = 2*self.size - self.x
-            self.angle = - self.angle
-
-        if self.y > dispHeight - self.size:
-            self.y = 2*(dispHeight - self.size) - self.y
-            self.angle = pi - self.angle
-
-        elif self.y < self.size:
-            self.y = 2*self.size - self.y
-            self.angle = pi - self.angle
-
-
-    def move(self, dist, angle, speed, smear=False):
-        self.angle = angle
-        self.speed = speed
-
-        # Adding boundary function :D
-        # s = "current distance of movement is " + str(dist)
-        # msg2screen(s,self.x - 100, self.y + 200)
-        while dist > 0:
-            # s = "current distance of movement is " + str(dist)
-            # msg2screen(s,self.x - 100, self.y + 200)
-            if not smear:
-                gameDisplay.fill(GREEN)
-
-            self.x += int(round(sin(self.angle) * self.speed))
-            self.y -= int(round(cos(self.angle) * self.speed))
-
-            #---------
-            self.boundary()
-            #---------
-
-            # if (self.x + self.size) > dispWidth or (self.x - self.size) < 0 or (self.y + self.size) > dispHeight or (self.y - self.size)< 0:
-            #     s = "I am breaking it bad .... :( " + str(dist)
-            #     msg2screen(s,self.x - 100, self.y + 200)
-            #     pygame.display.update()
-            #     break
-            # else:
-            #     s = "I am doing it good .... :) " + str(dist)
-            #     msg2screen(s,self.x - 100, self.y + 200)
-            #     pygame.display.update()
-
-            self.disp()
-            # pygame.display.update()
-            dist -= 1 # Movement of these units only :D
-        # s = "current distance of movement is " + str(dist)
-        # msg2screen(s,self.x - 100, self.y + 200)
-    def move_with_collision_detection(self,dist = None, angle = None, list_of_ball_objects = [], speed = 3, smear = False):
-        # NOTE: 
-        # This is specially used for the ball_point tracing :D
-        # Find the related code in trace_for_white_ball :D
-
-        # This variable is to detect the journey of ball_point
-        # If during its journey is is pocketed then, I will return 1, hence the white_ball
-        # can hit the shot, and ball will be pocketed :D
-        in_journey = 0
-
-        if dist != None:
-            self.dist = dist
-
-        if angle != None:
-            self.angle = angle
-
-        if list_of_ball_objects == []:
-            list_of_ball_objects.append(self)
-        
-        self.speed = speed
-
-        while self.dist > 0 & (self.pocketed == 0):
-            # s = "current distance of movement is " + str(dist)
-            # msg2screen(s,self.x - 100, self.y + 200)
-            if not smear:
-                gameDisplay.fill(GREEN)
-
-            self.x += int(round(sin(self.angle) * self.speed))
-            self.y -= int(round(cos(self.angle) * self.speed))
-            
-            p, q = self.x + 10, self.y + 10
-            s = str(self.dist)
-            msg2screen(s,p,q)
-            #---------
-            self.collision_detection(list_of_ball_objects, self.dist)
-
-            temp_ball_for_pocket_size = list_of_ball_objects[0]
-
-            temp = self.is_pocketed(temp_ball_for_pocket_size.pocket_size)
-
-            if temp == 1:
-                in_journey = 1
-                self.dist = 0       # If the ball_point is pocketed, then no more distance to move for it :D
-            else:
-                self.boundary()
-                #---------
-            self.disp()
-            pygame.display.update()
-            self.dist -= 1 # Movement of these units only :D
-        
-        return in_journey
-
-    def move_with_collision_correction(self,dist=None, angle = None, speed = None, list_of_ball_objects = []):
-        
-        if dist != None:
-            self.dist = dist
-
-        if angle != None:
-            self.angle = angle
-
-        if list_of_ball_objects == []:
-            list_of_ball_objects.append(self)
-        
-        if speed != None:
-            self.speed = speed
-
-        # Adding boundary function :D
-
-        # s = "current distance of movement is " + str(dist)
-        # msg2screen(s,self.x - 100, self.y + 200)
-        while self.dist > 0:
-            # s = "current distance of movement is " + str(dist)
-            # msg2screen(s,self.x - 100, self.y + 200)
-            
-            # gameDisplay.fill(GREEN)
-
-            self.x += int(round(sin(self.angle) * self.speed))
-            self.y -= int(round(cos(self.angle) * self.speed))
-            
-            # try:
-            #     self.x += int(round(sin(self.angle) * self.speed))
-            #     self.y -= int(round(cos(self.angle) * self.speed))
-            # except TypeError:
-            #     print "Self angle: " + str(self.angle) + "Self speed: " + str(self.speed) + "Type of self.speed: " + str(type(self.speed)) + "\n"
-            #     gameExit = True
-            #     break
-
-            #---------
-            self.collision(list_of_ball_objects, self.dist)
-            self.is_pocketed(self.pocket_size)
-            self.boundary()
-            #---------
-            self.disp()
-            # pygame.display.update()
-            self.dist -= 1 # Movement of these units only :D
-
-    def is_clicked(self, a, b):
-        if hypot(self.x-a, self.y-b) <= self.size:
-            return 1                                # If the point (a,b) is inside the ball then return 1 :D
-        else:
-            return 0                                # Else return 0 :D
-    
-    def collision_detection(self, temp_list_of_balls, dist):
-        # NOTE:
-        # Here I am only changing my self parameters, without changing anyone else parameters. Why???
-        # Ans:  This is used only ball_point, specially for white_ball collision tracing :D
-        #       Hence, I only want the trace of ball_point after hitting. I don't want to move the balls
-        #       if the ball_point trace hits them :D
-
-        # NOTE: temp_list_of_balls does not contain "self" by default, but it has white_ball :D
-        # Following step to check existence of self is for relality check :D
-
-        all_other_except_me = [my_ball for my_ball in temp_list_of_balls if (my_ball != self) & (my_ball.pocketed == 0)]
-        print "in collision_detection no. of balls came is " + str(len(all_other_except_me))
-        # This is list of all balls except me, which are not pocketed :D
-
-        for my_ball in all_other_except_me:
-            dx = my_ball.x - self.x
-            dy = my_ball.y - self.y
-
-            dist_of_separation = hypot(dx,dy)
-
-            if dist_of_separation < (self.size + my_ball.size):             # Collision happened :D
-                print "Bang Bang !! in collision detection"
-                
-                tangent = atan2(dy, dx)
-                self.angle      = 2*tangent - self.angle
-                self.dist       = self.dist/2
-
-                
-
-    def collision(self, all_balls, dist):
-        all_other_except_me = [my_ball for my_ball in all_balls if (my_ball != self) & (my_ball.pocketed == 0)]
-        # This is list of all balls except me, which are not pocketed :D
-
-        for my_ball in all_other_except_me:
-            dx = my_ball.x - self.x
-            dy = my_ball.y - self.y
-
-            dist_of_separation = hypot(dx,dy)
-
-            if dist_of_separation < (self.size + my_ball.size):             # Collision happened :D
-                print "Bang Bang !!"
-                
-                tangent = atan2(dy, dx)
-                self.angle      = 2*tangent - self.angle
-                self.dist       = self.dist/2
-
-                my_ball.angle   = 2*tangent - my_ball.angle
-                my_ball.dist    = self.dist             # This is modified distance, which 1/2 of the previous :D
-
-                # list_for_my_ball = [temp_ball for temp_ball in all_balls if temp_ball != my_ball]
-
-                my_ball.move_with_collision_correction(list_of_ball_objects = all_balls)
-
-    def is_pocketed(self, pocket_size):
-        x_limits = (0, dispWidth/2, dispWidth)
-        y_limits = (0, dispHeight)
-
-        for i in x_limits:
-            for j in y_limits:
-                ball_dist = hypot(self.x - i, self.y -j)
-                if ball_dist < pocket_size:
-                    self.pocketed = 1
-                    self.dist = 0           # So that the ball will not move any further :D
-                    return 1
-        return 0
-
-def create_rand_Ball(size):
-    x = random.randint(size, dispWidth-size)
-    y = random.randint(size, dispHeight-size)
-    return Balls((x,y), size)
-
-def msg2screen(msg, x_loc=dispWidth/4, y_loc=dispHeight/2, color=black):
-    screen_text = font.render(msg, True, color)
-    gameDisplay.blit(screen_text, [x_loc,  y_loc])
-
-def show_pockets(pocket_size):
-    a = (0, dispWidth/2, dispWidth)
-    b = (0, dispHeight)
-
-    for i in a:
-        for j in b:
-            pygame.draw.circle(gameDisplay, (0, 0, 0), (i, j), pocket_size, 0)
-            pygame.display.update()
-            
-def stick(block_size, stickList): # This will print our stick of finite length :D
-    for each in stickList:
-        pygame.draw.rect(gameDisplay, blue, [each[0], each[1], block_size, block_size])
-
-def get_points(x1, y1, x2, y2, curr_dist, m, choice):
-    # choice = 0 for start point , 
-    #        = 1 for end point :D
-
-    # In logic, (x1,y1) is start point, and (x2,y2) is end point
-    # What I am finding out it another point on the line at distance "curr_dist" in direction of movement :D
-
-    # if (y2 >= y1) ^ choice:
-    # # y2 = y2-1
-    #     y = y1 - curr_dist
-    # else:
-    # # y2 = y2+1
-    #     y = y1 + curr_dist
-
-    # x = x2 + round(m*(y - y2))  # From (x2,y2) and y selected, x will be found out :D
-    # a = dispWidth/4
-    # b = 3*dispHeight/4
-
-    if abs(y2-y1) >= abs(x2-x1):
-        if (y2 >= y1) ^ choice:
-        # y2 = y2-1
-            y = y1 - curr_dist
-        else:
-        # y2 = y2+1
-            y = y1 + curr_dist
-
-        x = x2 + round(m*(y - y2))  # From (x2,y2) and y selected, x will be found out :D
-        # s = "<<--------Inside y------->>  " + str(m)
-    else:
-        if (x2 >= x1) ^ choice:
-            x = x1 - curr_dist
-        else:
-            x = x1 + curr_dist
-
-        y = y2 + round(m*(x - x2))  # From (x2,y2) and y selected, x will be found out :D
-        # s = "<<--------Inside x------->>  " + str(m)
-
-    # msg2screen(s,a,b)
-    return (x, y)
-
-def new_get_slope(x1,y1,x2,y2):
-    # p = dispWidth/4
-    # q = 3*dispHeight/4
-    # s = "x1, y1, x2, y2: " + str(x1) + ":" + str(x1) + ":" + str(y1) + ":" + str(x2) + ":" + str(y2)
-
-    if (y2-y1)!=0:
-        m = float(x2-x1)/(y2-y1)
-    else:
-        m = abs(x2-x1)*10**2   # I am returing a high value, equivalent to infinity :D
-        # if y1 > y2:
-        #     m = m * -1
-        if x2 > x1:
-            m = m * -1
-    return m
-
-def get_slope(x1,y1,x2,y2):
-    # p = dispWidth/4
-    # q = 3*dispHeight/4
-    # s = "x1, y1, x2, y2: " + str(x1) + ":" + str(x1) + ":" + str(y1) + ":" + str(x2) + ":" + str(y2)
-    # # if (x2-x1)!=0:
-    #                                 #          (y2-y1)
-    #                 #                (y-y1) =  ------- * (x-x1)                                      
-    #                            #               (x2-x1) 
-    #     m1 = float(y2-y1)/(x2-x1)
-    #     m = m1
-    #     s = "you are is m111111111"
-    #     # return m1
-
-    # elif (y2-y1)!=0:
-    #                                 #          (x2-x1)
-    #                 #                (x-x1) =  ------- * (y-y1)                                      
-    #                            #               (y2-y1)                     
-    #     m2 = float(x2-x1)/(y2-y1)
-    #     m = m2
-    #     s = "you are is m222222222"
-    #     # return m2
-    # # msg2screen(s,p,q)
-    y_diff = (y2-y1)
-    x_diff = (x2-x1)
-
-    m = tan(atan2(y_diff,x_diff))
-    return m
-
-def get_angle(point_loc, Ball_obj):
-    temp1 = point_loc[0] - Ball_obj.x
-    temp2 = -1* (point_loc[1] - Ball_obj.y) # This -ve sign is to incorporate the inverted y-axis :D
-
-    # if temp2 == 0:
-    #     move_angle = (pi/2) * (temp1/abs(temp1)) # This is for the sign of than angle also :D
-    # else:
-    #     move_angle = atan(float(temp1)/temp2)
-
-    move_angle = atan2(temp1, temp2)
-
-    return move_angle
-
-def ball_got_hit(start_point, Ball_obj):
-    hamming_dist = sqrt((start_point[0]- Ball_obj.x)**2 + (start_point[1]- Ball_obj.y)**2)
-
-    if hamming_dist < Ball_obj.size:
-        return 1
-    else:
-        return 0
-def trace_the_shot(start_point, end_point, m, mouse_pos):
-    # This only projects a line from the stick with some finite distnace :D
-    # Ball touch is not included here :D
-
-    move_dist = sqrt((end_point[0] - mouse_pos[0])**2 + (end_point[1] - mouse_pos[1])**2)
-    # get_points(x1,y1,x2,y2,curr_dist,m,choice):
-    a,b = start_point[0], start_point[1], 
-    c,d = end_point[0], end_point[1]
-
-    # m = get_slope(a,b,c,d)
-    trace_point = get_points(a, b, c, d, move_dist, m, 0)  # Here choice = 0, Since I want to find the trace point :D
-
-    point_to_draw = []
-    point_to_draw.append(start_point)       # Line will start from here
-    # Now I am checking for any boundary reflection of that trace line :D
-
-    temp_dist = 1
-    # points_on_line = []
-
-    while temp_dist <= move_dist:
-        # m = get_slope(a,b,c,d)
-        # if abs(m)>1:
-        #     m = 1/m
-
-        temp_point = get_points(a, b, c, d, temp_dist, m, 0)  # Here choice = 0, Since I want to find the trace point :D
-        temp_dist += 1
-        # points_on_line.append(temp_point)
-
-        x,y = temp_point
-        # if x < 0 or y < 0 or x > (dispWidth ) or y > (dispHeight ):  
-        # # Here the condition is written in special way, so that point to added has positive cooradinates :D
-        #     last_point = points_on_line[-2]   
-        #     # I want to add the second last point which was added
-        #     # The last point which was added is the out of boundary point :D
-        #     point_to_draw.append(last_point)
-
-        #     # Modifying the "get_points" function call, so that line will be traced even after the boundary was crossed :D
-
-        #     # Checking which boundary was crossed
-        #     # 
-        #     #                                 this is boundary no. 1   (x---->--->)
-        #     #                               --------------------------
-        #     #                               |                        |
-        #     #                               |                        |
-        #     #     this is boundary no. 4    |                        |  this is boundary no. 2
-        #     #                               |                        |
-        #     #                               |                        |
-        #     #                               --------------------------
-        #     #                                   this is boundary no. 3
-
-        #     if x < 0 or x > (dispWidth ):
-        #         x = dispWidth - (x % dispWidth)
-
-        #     if y < 0 or y > (dispHeight ):
-        #         y = dispHeight - (y % dispHeight)
-
-        #     # Now after solving boundary problem, we need to modify the start and end points so that the line is traced correct
-        #     # Hence, "new end point" is the boundary point that added
-        #     # and "new start point" will be the out of boundary point which has been corrected as above :D
-        #     c,d = last_point        # new end point
-        #     a,b = x,y               # new start point :D
-    
-        k = 0 
-        # This is special unit :D
-
-        if x < k or y < k or x > (dispWidth -k) or y > (dispHeight -k):  
-            
-            # Here the condition is written in special way, so that point to added has positive cooradinates :D
-            
-            # Modifying the "get_points" function call, so that line will be traced even after the boundary was crossed :D
-
-            # Checking which boundary was crossed
-            # 
-            #                                 this is boundary no. 1   (x---->--->)
-            #                               --------------------------
-            #                               |                        |
-            #                               |                        |
-            #     this is boundary no. 4    |                        |  this is boundary no. 2
-            #                               |                        |
-            #                               |                        |
-            #                               --------------------------
-            #                                   this is boundary no. 3
-            
-            
-
-            if x < k or x > (dispWidth - k):
-                x = dispWidth - (x % dispWidth)
-                if x < k:
-                    x_cordi = 0
-
-                    c = c
-                    d = -d 
-                    
-                else:
-                    x_cordi = dispWidth
-
-                    c = c
-                    d = 2*dispHeight - d
-
-                y_cordi = m*x_cordi + (b - m*a)  
-
-            if y < k or y > (dispHeight -k):
-                y = dispHeight - (y % dispHeight)
-                if y < k:
-                    y_cordi = 0
-
-                    c = -c
-                    d = d 
-
-                else:
-                    y_cordi = dispHeight
-
-                    c = 2*dispWidth - c
-                    d = d
-
-                x_cordi = (1/m)*y_cordi + (a - (1/m)*b)  # You can replace y2 by y2 , and x2 by x1 also :D
-
-            last_point = (x_cordi, y_cordi)
-            point_to_draw.append(last_point)
-
-            # Now after solving boundary problem, we need to modify the start and end points so that the line is traced correct
-            # Hence, "new start point" is the boundary point that added
-            # and "new end point" will be calculated as below (I have reflected previous end point to get this new one :D)
-            a,b = last_point        # new start point
-                                    # new start point :D already modified
-         # After finishing the while loop, I must add the last point which got traced, to list of "point_to_draw"       
-
-        m = tan(pi - atan(m))  # AFter coming out of the loop for correcting x,y change the slop of the line :D
-        if abs(m)>1:
-            m = 1/m
-
-
-    point_to_draw.append(temp_point)
-
-     # I will now draw lines in between the point in my list "point_to_draw"
-     # i.e. line between 1st and 2nd, then between 2nd and 3rd, and so on
-     # Why/when "point_to_draw" will have more than 2 points??
-     # If there is reflection due to surface then only more than 2
-     # If no reflection at all then only 2 point in the list "point_to_draw" :D
-    print point_to_draw
-    for i in xrange(len(point_to_draw) - 1):  
-     # Ques: Why    len(point_to_draw) -1    ? 
-     # Ans : Because to no pair for the last point :D
-        # pygame.draw.aaline(gameDisplay, (0, 255, 0), point_to_draw[i], point_to_draw[i+1])
-        # pygame.draw.line(gameDisplay, (255, 0, 0), point_to_draw[i], point_to_draw[i+1],20)
-        pygame.display.update()
-    aa,bb = trace_point
-    # aa = aa % dispWidth
-    # bb = bb % dispHeight
-    new = aa,bb
-    pygame.draw.line(gameDisplay, (255, 0, 0), point_to_draw[0], new,2)
-    pygame.display.update()
-    # p = dispWidth/4 - 100
-    # q = 3*dispHeight/4 + 100
-    # s = "tracing the shot.... No of lines to be drawn is : "+ str(len(point_to_draw) -1) +"::: " + str(point_to_draw)
-    # msg2screen(s,p,q)
-    # pygame.display.update()
-
-def in_boundary(trace_point):
-    a,b = trace_point
-    if a < 0 or b < 0 or a > dispWidth or b > dispHeight:
-        return 0
-    else:
-        return 1
-
-# def get_boundary_point(trace_point, m): # This function will be called only if the point is out of boundary :D
-#     a,b = trace_point
-def get_elevation(start_point, end_point):
-    a,b = start_point
-    c,d = end_point
-    new_m = new_get_slope(a,b,c,d)  # Calculating this slope :D
-    
-    if new_m > 0:   # First and Third quadrant
-        if b < d:
-            elevation = pi + atan(new_m)
-        else:
-            elevation = atan(new_m)
-    else:           # Second and Fourth quadrant
-        if b < d:
-            elevation = pi + atan(new_m)
-        else:
-            elevation = 2*pi + atan(new_m)
-
-    return elevation
-
-def trace_for_while_ball_shot(my_ball_loc, white_ball_loc, list_of_ball_objects, move_dist = 50 ):
-    
-    elevation = get_elevation(my_ball_loc, white_ball_loc)
-
-    # Balls(self,(x,y),size, thickness=0, color=(0,0,255)):
-    ball_point = Balls(my_ball_loc, 3, color=(255,0,0))
-    
-    # kkk = 2*pi/2
-    kkk = pi
-    slope_angle = elevation
-    trace_angle = kkk - (slope_angle)
-
-    # Note: The method is used is different :D
-    # Its detection only :D no correction is intended since its only trace :D
-    in_journey = ball_point.move_with_collision_detection(move_dist, trace_angle, list_of_ball_objects, speed = 8, smear=True )
-                            
-    pygame.display.update()
-
-    will_it_be_pocketed = in_journey            # yes for 1, and no for 0 :D
-
-    return will_it_be_pocketed
-
-def new_trace_the_shot(start_point, end_point, elevation, mouse_pos):
-    a,b = end_point
-    c,d = mouse_pos
-    e,f = start_point
-    
-    move_dist       = sqrt((a-c)**2 + (b-d)**2)
-
-    # Balls(self,(x,y),size, thickness=0, color=(0,0,255)):
-    ball_point = Balls(start_point, 3, color=(255,0,0))
-    
-    # kkk = 2*pi/2
-    kkk = pi
-    slope_angle = elevation
-    trace_angle = kkk - (slope_angle)
-
-    # Ball.move(self,dist, angle, speed)
-    ball_point.move(move_dist, trace_angle, speed = 8, smear=True)
-
-    # p = dispWidth/4
-    # q = 3*dispHeight/4
-    # s = "tracing the shot....at angle : " + str(trace_angle*180/pi)
-    s = "tracing the shot...................(" + str(slope_angle*180/pi) +").........(" + str(trace_angle*180/pi) + ")"
-    print s
-    # msg2screen(s,p,q)
-    # pygame.display.update()
-
+# #============
+# # Functions are written to some py file :D
+# from modules.all_functions import *
+# #============
 def gameLoop():
     gameExit = False
     gameOver = False
@@ -678,12 +59,6 @@ def gameLoop():
 
         all_balls.append(Balls((x,y), size = my_ball_size, color = (c1,c2,c3)))
     
-    # my_pocket_size = 2*my_ball_size
-    # show_pockets(my_pocket_size)
-    # pygame.display.update()
-    # Ball_1 = Balls((320,400),15)            # Fixed location ball
-                                            # Initialising the ball here :D
-
     print "++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++"
     # screen = pygame.display.set_mode((dispHeight, dispWidth), 0, 32)
     while not gameExit:
@@ -839,7 +214,13 @@ def gameLoop():
                     list_of_balls_with_white = [a_ball for a_ball in all_balls]
                     list_of_balls_with_white.append(white_ball)
 
-                    white_ball.move_with_collision_correction(dist = rand_dist, angle = move_angle, list_of_ball_objects = list_of_balls_with_white)
+                    # white_ball.move_with_collision_correction(dist = rand_dist, angle = move_angle, list_of_ball_objects = list_of_balls_with_white)
+
+                    white_ball.dist = rand_dist
+                    white_ball.angle = move_angle
+
+                    move_my_all_balls(list_of_balls_with_white)
+
                     at_least_one_ball_hit = 1               # 1 is for YES, on ball got hit, hence breaking :D
 
                     p, q = hit_ball.x + 30, hit_ball.y + 30
@@ -1062,7 +443,12 @@ def gameLoop():
                                 move_angle = temp_angle + pi                # Why this??? Ball should move in opposite direction from where it is being hit :D
                                 move_speed = 3                              # You can choose more high speed :D
 
-                                my_ball.move_with_collision_correction(dist = ball_move_dist, angle = move_angle, speed = move_speed, list_of_ball_objects = all_balls_with_white)
+                                # my_ball.move_with_collision_correction(dist = ball_move_dist, angle = move_angle, speed = move_speed, list_of_ball_objects = all_balls_with_white)
+                                my_ball.dist = ball_move_dist
+                                my_ball.angle = move_angle
+                                my_ball.speed = move_speed
+
+                                move_my_all_balls(all_balls_with_white)
                                 # move_with_collision_correction(self,dist=None, angle = None, speed=3, list_of_ball_objects =[]):
                                 # pygame.display.update()
